Move HIR traversal queries in rustc_ast_lowering.
